refactor: extract edgesentry-types crate; move geo math to edgesentry-compute — closes #337
- New crate edgesentry-types: Entity, Vec2, EntityClass (no I/O, no deps)
- edgesentry-compute depends on edgesentry-types only — no longer depends on edgesentry-ingest
- edgesentry-compute/src/geo.rs: latlon_to_local + cog_sog_to_velocity moved from ais_nmea.rs
- edgesentry-ingest/src/entity.rs: replaced with pub use edgesentry_types::* (backward compat)
- edgesentry-ingest/src/ais_nmea.rs: geo math removed, now uses edgesentry-compute
- All workspace crates updated: evaluate, parse, eds
- All tests pass; clarus cargo check clean
