+++ crates/edgesentry-compute/Cargo.toml
@@ -7,6 +7,6 @@ repository.workspace = true
 description = "Physics and geometry computations — distance, TTC, zone membership"
 
 [dependencies]
-edgesentry-ingest = { path = "../edgesentry-ingest" }
+edgesentry-types = { path = "../edgesentry-types" }
 serde = { workspace = true }
 serde_json = { workspace = true }
