+++ crates/edgesentry-compute/src/lib.rs
@@ -1,2 +1,4 @@
 pub mod physics;
+pub mod geo;
 pub use physics::*;
+pub use geo::{latlon_to_local, cog_sog_to_velocity};
