+++ crates/edgesentry-compute/src/physics.rs
@@ -1,4 +1,4 @@
-use edgesentry_ingest::entity::{Entity, EntityClass, Vec2};
+use edgesentry_types::{Entity, EntityClass, Vec2};
 
 /// Straight-line distance between two entities in metres.
 pub fn euclidean_distance(a: &Entity, b: &Entity) -> f32 {
@@ -74,7 +74,7 @@ pub fn zone_membership(pos: Vec2, polygon: &[Vec2]) -> bool {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use edgesentry_ingest::entity::{Entity, EntityClass, Vec2};
+    use edgesentry_types::{Entity, EntityClass, Vec2};
 
     fn entity(x: f32, y: f32, vx: f32, vy: f32) -> Entity {
         Entity {
