+++ crates/edgesentry-parse/Cargo.toml
@@ -8,6 +8,7 @@ description = "Maritime structured data parsing — CSV/Parquet → DocumentEnti
 
 [dependencies]
 edgesentry-ingest = { path = "../edgesentry-ingest" }
+edgesentry-types = { path = "../edgesentry-types" }
 serde = { workspace = true }
 serde_json = { workspace = true }
 parquet = { version = "55", default-features = false, features = ["snap"] }
